refactor: split protocol_test.go into per-concern test files
The monolithic protocol_test.go (1259 lines, 17 tests) was becoming a
merge-conflict magnet as multiple PRs add tests. Split into:

- helpers_test.go: shared infrastructure (workspace, build, parse, assert)
- import_test.go: import/export round-trip fidelity (2 tests)
- roundtrip_test.go: field/metadata/spec/close-reason round-trips (4 tests)
- integrity_test.go: data integrity under mutations (5 tests)
- ready_test.go: ready ordering and blocking semantics (5 tests)

All tests share the workspace helpers via Go's same-package test mechanism.
New PRs add tests to the appropriate file without touching unrelated tests.

Also adds workspace.tryRun() helper for tests that expect command failure
(needed by upcoming validation fix PRs).
